Organize server dir
