+++ server/scripts/seed.ts
@@ -3,10 +3,10 @@ import { exit } from 'process'
 import faker from 'faker'
 
 import mongoose from 'mongoose'
-import { connectDB } from './mongoConnection'
-import { Request } from './models/requestModel'
-import { RequestGroup } from './models/requestGroupModel'
-import { RequestType } from './models/requestTypeModel'
+import { connectDB } from '../database/mongoConnection'
+import { Request } from '../models/requestModel'
+import { RequestGroup } from '../models/requestGroupModel'
+import { RequestType } from '../models/requestTypeModel'
 
 // -----------------------------------------------------------------------------
 // SEED REQUESTS/TAGS
